pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trash-talk-go-bot"
version = "0.1.0"
description = "A Go-playing bot with personality that uses KataGo via MCP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=0.2.0",
    "python-ai-sdk>=0.1.0",
    "structlog>=24.2.0",
    "pyyaml>=6.0.0",
    "asyncio>=3.4.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]
